requirements.txt
# Minimum version of pillow library
pillow >= 11.1.0